Code cleanup Fixed textinput not focusing
